+++ src/utils/Settings.ts
@@ -24,8 +24,8 @@ export class Settings {
   static getSettings(val: string) {
     return Settings.configuration.get(val);
   }
-  static setSettings(key: string, val: any, isGlobal = true) {
-    return Settings.configuration.update(key, val, isGlobal);
+  static setSettings(key: string, val: any) {
+    return Settings.configuration.update(key, val, vscode.ConfigurationTarget.Global);
   }
 
   static get paths() {
@@ -47,22 +47,40 @@ export class Settings {
     const defaultInclude: string[] = [];
     const defaultExclude: string[] = ['node_modules', 'dist', 'build', 'out'];
 
+    const interpolateObject = {
+      workspaceFolder: workspaceFolders[0]?.uri.fsPath || '',
+      userHome,
+      workspaceFolderName: workspaceFolders[0]?.name || '',
+      workspaceFolderBasename: path.basename(workspaceFolders[0]?.uri.fsPath || ''),
+    };
+
+    function extractVariableAndValue(input: string) {
+      // Match ${variableName: value} followed by optional suffix
+      const match = input.match(/\$\{([^:}]+):\s*([^}]*)\}([^\s]*)?/);
+      if (!match) return null;
+
+      const variableWithSuffix = `\${${match[1].trim()}}${match[3] || ''}`;
+      const value = match[2].trim();
+      return [variableWithSuffix, value];
+    }
+
     // get Normalized paths
     const normalized: NormalizedPaths[] = paths.map((p) => {
       if (typeof p === 'string') {
-        const basePath = interpolate(p.replace(/\\/g, '/'), { workspaceFolder: workspaceFolders[0]?.uri.fsPath || '', userHome });
+        const [variable = p, folderName] = extractVariableAndValue(p || '${workspaceFolder}') || [];
+        const basePath = interpolate(variable.replace(/\\/g, '/') || '${workspaceFolder}', interpolateObject);
+        const resolvedBasePath = path.resolve(interpolateObject.workspaceFolder, basePath); // resolve with workspace folder to support multiple folders
         return {
-          basePath,
-          name: path.basename(basePath),
+          basePath: resolvedBasePath,
+          name: folderName || path.basename(resolvedBasePath),
         };
       }
-      const basePath = interpolate(p.basePath?.replace(/\\/g, '/') || `${workspaceFolders}`, {
-        workspaceFolder: workspaceFolders[0]?.uri.fsPath || '',
-        userHome,
-      });
+      const [variable = p.basePath, folderName] = extractVariableAndValue(p.basePath || '${workspaceFolder}') || [];
+      const basePath = interpolate(variable?.replace(/\\/g, '/') || '${workspaceFolder}', interpolateObject);
+      const resolvedBasePath = path.resolve(interpolateObject.workspaceFolder, basePath); // resolve with workspace folder to support multiple folders
       return {
-        basePath,
-        name: p.name || path.basename(basePath),
+        basePath: resolvedBasePath,
+        name: interpolate(p.name || folderName || path.basename(resolvedBasePath), interpolateObject),
         include: getFormattedPatternPaths(([] as string[]).concat(p.include || defaultInclude)),
         exclude: getFormattedPatternPaths(([] as string[]).concat(p.exclude || defaultExclude)),
       };
@@ -70,7 +88,6 @@ export class Settings {
 
     // filter only valid and exist baseFolders
     const filtered = normalized.filter((p) => p.basePath && path.isAbsolute(p.basePath) && fsx.existsSync(p.basePath));
-    console.log('Filtered paths:', filtered);
     return filtered;
   }
 }